Ball module

+++ ball.h
@@ -1,12 +1,24 @@
+#include <stdint.h>
+
+#include "paddle.h"
+
+// ball structure with movement variables
 struct ball;
 typedef struct ball ball;
 
+// creates a new ball in the center of the grid
 ball *newBall (grid *g, paddle *p);
 
-uint8_t getBallX (ball *b);
-uint8_t getBallY (ball *b);
+// returns the x pos. of the ball
+float getBallX (ball *b);
+// returns the y pos. of the ball
+float getBallY (ball *b);
 
-uint8_t getBallVelX (ball *b);
-uint8_t getBallVelY (ball *b);
+// returns the horiz. vel of the ball
+float getBallVelX (ball *b);
+// returns the vert. vel. of the ball
+float getBallVelY (ball *b);
 
+// changes the pos. of the ball
+// according to it's velocity & collisions
 void tick ();
